backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pillow
python-magic
exifread
ffmpeg-python
taskiq[redis]
# AI 相关 (按需启用)
# insightface
# onnxruntime
# sentence-transformers
